backend/pyproject.toml
[project]
name = "exam-backend"
version = "1.0.0"
requires-python = ">=3.10"

dependencies = [
    "fastapi",
    "uvicorn",
    "psycopg[binary]",
    "python-dotenv"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-bdd",
    "httpx"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
